fixed some printing bugs
previously, if a width specifier was used and the next argument to
kprintf wasn't zero padded it would erroneously insert a very large
number of spaces. This has been fixed.

In addition, tabs would not correctly wrap at the ends of lines. this
has also been fixed.

the virtual to physical translation function now correctly supports
virtual addresses with the first 16 bits set to 1, such as those the
kernel uses.